Crate posei_traderbento

PoseiTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

This crate provides a Databento integration adapter.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the posei_trader Python package, or as part of a Rust only build.

python: Enables Python bindings from PyO3.

Modules

common

Common functions to support Databento adapter operations.

data

Databento data client implementation leveraging existing live and historical clients.

decode

enums

Enumerations for the Databento integration.

factories

Factory functions for creating Databento clients and components.

historical

Core Databento historical client for both Rust and Python usage.

live

loader

python

Python bindings from PyO3.

symbology

types