posei_traderbento::actor

Struct DatabentoSubscriberActor

pub struct DatabentoSubscriberActor {
    pub received_quotes: Vec<QuoteTick>,
    pub received_trades: Vec<TradeTick>,
    /* private fields */
}

A basic Databento subscriber actor that subscribes to quotes and trades.

This actor demonstrates how to use the DataActor trait to subscribe to market data from Databento for specified instruments. It logs received quotes and trades to demonstrate the data flow.

Fields
received_quotes: Vec<QuoteTick>received_trades: Vec<TradeTick>
Implementations

impl DatabentoSubscriberActor

pub fn new( config: DatabentoSubscriberActorConfig, cache: Rc<RefCell<Cache>>, clock: Rc<RefCell<dyn Clock>>, ) -> Self

Creates a new DatabentoSubscriberActor instance.

pub fn quote_count(&self) -> usize

Returns the number of quotes received by this actor.

pub fn trade_count(&self) -> usize

Returns the number of trades received by this actor.