posei_traderbento

Module decode

Functions

decode_bar_type

Errors

decode_bbo_msg

Errors

decode_equity

Errors

decode_futures_contract

Errors

decode_futures_spread

Errors

decode_imbalance_msg

Errors

decode_instrument_def_msg

Errors

decode_lot_size

Decodes a lot size from the given value, expressed in standard whole-number units.

decode_mbo_msg

Errors

decode_mbp1_msg

Errors

decode_mbp10_msg

Errors

decode_multiplier

Decodes a multiplier from the given value, expressed in units of 1e-9.

decode_ohlcv_msg

Errors

decode_option_contract

Errors

decode_option_spread

Errors

decode_optional_price

Decodes a price from the given optional value, expressed in units of 1e-9.

decode_optional_quantity

Decodes a quantity from the given optional value, expressed in standard whole-number units.

decode_price

Decodes a price from the given value, expressed in units of 1e-9.

decode_price_increment

Decodes a minimum price increment from the given value, expressed in units of 1e-9.

decode_quantity

Decodes a quantity from the given value, expressed in standard whole-number units.

decode_record

Errors

decode_statistics_msg

Errors

decode_status_msg

Errors

decode_tbbo_msg

Errors

decode_trade_msg

Errors

decode_ts_event_adjustment

Errors

parse_aggressor_side

parse_book_action

Errors

parse_cfi_iso10926

Errors

parse_option_kind

Errors

parse_optional_bool

parse_order_side

parse_status_reason

Errors

parse_status_trading_event

Errors