posei_traderbento::loader

Struct DatabentoDataLoader

pub struct DatabentoDataLoader { /* private fields */ }

A Posei data loader for Databento Binary Encoding (DBN) format data.

Supported schemas:

MBO -> OrderBookDelta

MBP_1 -> (QuoteTick, Option<TradeTick>)

MBP_10 -> OrderBookDepth10

BBO_1S -> QuoteTick

BBO_1M -> QuoteTick

TBBO -> (QuoteTick, TradeTick)

TRADES -> TradeTick

OHLCV_1S -> Bar

OHLCV_1M -> Bar

OHLCV_1H -> Bar

OHLCV_1D -> Bar

DEFINITION -> Instrument

IMBALANCE -> DatabentoImbalance

STATISTICS -> DatabentoStatistics

STATUS -> InstrumentStatus

References

https://databento.com/docs/schemas-and-data-formats

Implementations

impl DatabentoDataLoader

pub fn new(publishers_filepath: Option<PathBuf>) -> Result<Self>

Creates a new DatabentoDataLoader instance.

Errors

Returns an error if locating or loading publishers data fails.

pub fn load_publishers(&mut self, filepath: PathBuf) -> Result<()>

Load the publishers data from the file at the given filepath.

Errors

Returns an error if the file cannot be read or parsed as JSON.

pub const fn get_publishers(&self) -> &IndexMap<u16, DatabentoPublisher>

Returns the internal Databento publishers currently held by the loader.

pub fn set_dataset_for_venue(&mut self, dataset: Dataset, venue: Venue)

Sets the venue to map to the given dataset.

pub fn get_dataset_for_venue(&self, venue: &Venue) -> Option<&Dataset>

Returns the dataset which matches the given venue (if found).

pub fn get_venue_for_publisher( &self, publisher_id: PublisherId, ) -> Option<&Venue>

Returns the venue which matches the given publisher_id (if found).

pub fn schema_from_file(&self, filepath: &Path) -> Result<Option<String>>

Returns the schema for the given filepath.

Errors

Returns an error if the file cannot be decoded or metadata retrieval fails.

pub fn read_definition_records( &mut self, filepath: &Path, use_exchange_as_venue: bool, ) -> Result<impl Iterator<Item = Result<InstrumentAny>> + '_>

Errors

Returns an error if decoding the definition records fails.

pub fn read_records<T>( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, include_trades: bool, bars_timestamp_on_close: Option<bool>, ) -> Result<impl Iterator<Item = Result<(Option<Data>, Option<Data>)>> + '_>
where T: Record + HasRType + 'static,

Errors

Returns an error if reading records fails.

pub fn load_instruments( &mut self, filepath: &Path, use_exchange_as_venue: bool, ) -> Result<Vec<InstrumentAny>>

Errors

Returns an error if loading instruments fails.

pub fn load_order_book_deltas( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<Vec<OrderBookDelta>>

Errors

Returns an error if loading order book deltas fails.

pub fn load_order_book_depth10( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<Vec<OrderBookDepth10>>

Errors

Returns an error if loading order book depth10 fails.

pub fn load_quotes( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<Vec<QuoteTick>>

Errors

Returns an error if loading quotes fails.

pub fn load_bbo_quotes( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<Vec<QuoteTick>>

Errors

Returns an error if loading BBO quotes fails.

pub fn load_tbbo_trades( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<Vec<TradeTick>>

Errors

Returns an error if loading TBBO trades fails.

pub fn load_trades( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<Vec<TradeTick>>

Errors

Returns an error if loading trades fails.

pub fn load_bars( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, timestamp_on_close: Option<bool>, ) -> Result<Vec<Bar>>

Errors

Returns an error if loading bars fails.

pub fn load_status_records<T>( &self, filepath: &Path, instrument_id: Option<InstrumentId>, ) -> Result<impl Iterator<Item = Result<InstrumentStatus>> + '_>
where T: Record + HasRType + 'static,

Errors

Returns an error if loading status records fails.

pub fn read_imbalance_records<T>( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<impl Iterator<Item = Result<DatabentoImbalance>> + '_>
where T: Record + HasRType + 'static,

Errors

Returns an error if reading imbalance records fails.

pub fn read_statistics_records<T>( &self, filepath: &Path, instrument_id: Option<InstrumentId>, price_precision: Option<u8>, ) -> Result<impl Iterator<Item = Result<DatabentoStatistics>> + '_>
where T: Record + HasRType + 'static,

Errors

Returns an error if reading statistics records fails.